+++ internal/service/user_service.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/go-eagle/eagle/internal/dao"
+	"github.com/go-eagle/eagle/internal/repository"
 
 	"github.com/pkg/errors"
 
@@ -33,13 +33,13 @@ type UserService interface {
 }
 
 type userService struct {
-	dao *dao.Dao
+	repo repository.Repository
 }
 
 var _ UserService = (*userService)(nil)
 
 func newUsers(svc *service) *userService {
-	return &userService{dao: svc.dao}
+	return &userService{repo: svc.repo}
 }
 
 // Register 注册用户
@@ -56,14 +56,14 @@ func (s *userService) Register(ctx context.Context, username, email, password st
 		CreatedAt: time.Time{},
 		UpdatedAt: time.Time{},
 	}
-	isExist, err := s.dao.UserIsExist(&u)
+	isExist, err := s.repo.UserIsExist(&u)
 	if err != nil {
 		return errors.Wrapf(err, "create user")
 	}
 	if isExist {
 		return errors.New("用户已存在")
 	}
-	_, err = s.dao.CreateUser(ctx, &u)
+	_, err = s.repo.CreateUser(ctx, &u)
 	if err != nil {
 		return errors.Wrapf(err, "create user")
 	}
@@ -119,7 +119,7 @@ func (s *userService) PhoneLogin(ctx context.Context, phone int64, verifyCode in
 			Phone:    phone,
 			Username: strconv.Itoa(int(phone)),
 		}
-		u.ID, err = s.dao.CreateUser(ctx, &u)
+		u.ID, err = s.repo.CreateUser(ctx, &u)
 		if err != nil {
 			return "", errors.Wrapf(err, "[login] create user err")
 		}
@@ -136,7 +136,7 @@ func (s *userService) PhoneLogin(ctx context.Context, phone int64, verifyCode in
 
 // UpdateUser update user info
 func (s *userService) UpdateUser(ctx context.Context, id uint64, userMap map[string]interface{}) error {
-	err := s.dao.UpdateUser(ctx, id, userMap)
+	err := s.repo.UpdateUser(ctx, id, userMap)
 
 	if err != nil {
 		return err
@@ -147,12 +147,12 @@ func (s *userService) UpdateUser(ctx context.Context, id uint64, userMap map[str
 
 // GetUserByID 获取单条用户信息
 func (s *userService) GetUserByID(ctx context.Context, id uint64) (*model.UserBaseModel, error) {
-	return s.dao.GetOneUser(ctx, id)
+	return s.repo.GetUser(ctx, id)
 }
 
 // GetUserByID 获取单条用户信息
 func (s *userService) GetUserStatByID(ctx context.Context, id uint64) (*model.UserStatModel, error) {
-	return s.dao.GetUserStatByID(ctx, id)
+	return s.repo.GetUserStatByID(ctx, id)
 }
 
 // GetUserInfoByID 获取组装好的用户数据
@@ -170,13 +170,13 @@ func (s *userService) GetUserInfoByID(ctx context.Context, id uint64) (*model.Us
 func (s *userService) BatchGetUsers(ctx context.Context, userID uint64, userIDs []uint64) ([]*model.UserInfo, error) {
 	infos := make([]*model.UserInfo, 0)
 	// 获取当前用户信息
-	curUser, err := s.dao.GetOneUser(ctx, userID)
+	curUser, err := s.repo.GetUser(ctx, userID)
 	if err != nil {
 		return nil, errors.Wrap(err, "[user_service] get one user err")
 	}
 
 	// 批量获取用户信息
-	users, err := s.dao.GetUsersByIds(ctx, userIDs)
+	users, err := s.repo.GetUsersByIds(ctx, userIDs)
 	if err != nil {
 		return nil, errors.Wrap(err, "[user_service] batch get user err")
 	}
@@ -187,19 +187,19 @@ func (s *userService) BatchGetUsers(ctx context.Context, userID uint64, userIDs 
 	finished := make(chan bool, 1)
 
 	// 获取自己对关注列表的关注状态
-	userFollowMap, err := s.dao.GetFollowByUIds(ctx, userID, userIDs)
+	userFollowMap, err := s.repo.GetFollowByUIds(ctx, userID, userIDs)
 	if err != nil {
 		errChan <- err
 	}
 
 	// 获取自己对关注列表的被关注状态
-	userFansMap, err := s.dao.GetFansByUIds(ctx, userID, userIDs)
+	userFansMap, err := s.repo.GetFansByUIds(ctx, userID, userIDs)
 	if err != nil {
 		errChan <- err
 	}
 
 	// 获取用户统计
-	userStatMap, err := s.dao.GetUserStatByIDs(ctx, userIDs)
+	userStatMap, err := s.repo.GetUserStatByIDs(ctx, userIDs)
 	if err != nil {
 		errChan <- err
 	}
@@ -269,7 +269,7 @@ func (s *userService) BatchGetUsers(ctx context.Context, userID uint64, userIDs 
 }
 
 func (s *userService) GetUserByPhone(ctx context.Context, phone int64) (*model.UserBaseModel, error) {
-	userModel, err := s.dao.GetUserByPhone(ctx, phone)
+	userModel, err := s.repo.GetUserByPhone(ctx, phone)
 	if err != nil {
 		return userModel, errors.Wrapf(err, "get user info err from db by phone: %d", phone)
 	}
@@ -278,7 +278,7 @@ func (s *userService) GetUserByPhone(ctx context.Context, phone int64) (*model.U
 }
 
 func (s *userService) GetUserByEmail(ctx context.Context, email string) (*model.UserBaseModel, error) {
-	userModel, err := s.dao.GetUserByEmail(ctx, email)
+	userModel, err := s.repo.GetUserByEmail(ctx, email)
 	if err != nil {
 		return userModel, errors.Wrapf(err, "get user info err from db by email: %s", email)
 	}