+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## Changelog
 
-## v1.3.2
-- 新增 run 命令工具
+## v1.5.0
+- add: cache, repo, service cmd
+- add: grpc error
+- optimize: proto cmd
+
+## v1.4.0
+- 新增 run,proto,protoc-gen-go-gin 命令工具
 - 新增 服务注册
+- 新增 examples
 - 优化 replace OpenTracing with OpenTelemetry
 - 优化 add trace for http client and sql